Got the form mostly completed. Parents and Spouses now save correctly.
I need to test this with 'new' individuals still. Also, there should be some tests done for editing different individuals, with full content.
The next big step is to do the file upload directive so that I can insert it into the individual form (that way you can add a profile picture)